Atualizado

+++ real para dolar.py
@@ -1,6 +1,6 @@
-valor = int(input("informe o valor  do produto em reais: "))
+valor = float(input("informe o valor  do produto em reais: "))
 valordolar = float(input("Qual o valor do dolar hoje?: "))
-cotacaodolar = valor/valordolar
+cotacaodolar = float(valor/valordolar)
 
 print ("O valor em reais é: ", valor)
 print (f"O Valor em dolar é:  {cotacaodolar :.2f}")
